+++ axum/src/routing/mod.rs
@@ -1,6 +1,10 @@
 //! Routing between [`Service`]s and handlers.
 
-use self::{future::RouteFuture, not_found::NotFound};
+use self::{
+    future::RouteFuture,
+    inner_routes::{InnerRoutes, InsertError, Match},
+    not_found::NotFound,
+};
 use crate::{
     body::{boxed, Body, Bytes, HttpBody},
     extract::connect_info::IntoMakeServiceWithConnectInfo,
@@ -10,13 +14,10 @@ use crate::{
     BoxError,
 };
 use http::Request;
-use matchit::MatchError;
 use std::{
     borrow::Cow,
-    collections::HashMap,
     convert::Infallible,
     fmt,
-    sync::Arc,
     task::{Context, Poll},
 };
 use tower::{layer::layer_fn, ServiceBuilder};
@@ -26,6 +27,7 @@ use tower_service::Service;
 
 pub mod future;
 
+mod inner_routes;
 mod into_make_service;
 mod method_filter;
 mod method_routing;
@@ -45,22 +47,6 @@ pub use self::method_routing::{
     trace_service, MethodRouter,
 };
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
-struct RouteId(u32);
-
-impl RouteId {
-    fn next() -> Self {
-        use std::sync::atomic::{AtomicU32, Ordering};
-        // `AtomicU64` isn't supported on all platforms
-        static ID: AtomicU32 = AtomicU32::new(0);
-        let id = ID.fetch_add(1, Ordering::Relaxed);
-        if id == u32::MAX {
-            panic!("Over `u32::MAX` routes created. If you need this, please file an issue.");
-        }
-        Self(id)
-    }
-}
-
 pub(crate) const PRIVATE_PARAM_PREFIX: &str = "__private__axum";
 pub(crate) const NEST_TAIL_PARAM: &str = "__private__axum_nest_tail_param";
 pub(crate) const NEST_TAIL_PARAM_CAPTURE: &str = "/*__private__axum_nest_tail_param";
@@ -68,8 +54,14 @@ pub(crate) const NEST_FALLBACK_PARAM_CAPTURE: &str = "/*__private__axum_fallback
 
 /// The router type for composing handlers and services.
 pub struct Router<B = Body> {
+    /// Routes added with `Router::route`
     routes: InnerRoutes<Endpoint<B>>,
+    /// Custom fallbacks added with `Router::fallback`
+    ///
+    /// These get their own route because of fallbacks on nested routers.
     custom_fallbacks: InnerRoutes<Route<B>>,
+    /// The default fallbacks. We need to store these so we have something
+    /// to apply layers to in `Router::layer`.
     default_fallback: DefaultFallback<B>,
 }
 
@@ -110,27 +102,13 @@ where
 
         let endpoint = match try_downcast::<MethodRouter<B, Infallible>, _>(service) {
             Ok(method_router) => {
-                let prev_id_and_endpoint =
-                    self.routes
-                        .node
-                        .path_to_route_id
-                        .get(path)
-                        .and_then(|route_id| {
-                            self.routes
-                                .route_id_to_endpoint
-                                .get(route_id)
-                                .map(|svc| (*route_id, svc))
-                        });
-
-                if let Some((route_id, Endpoint::MethodRouter(prev_method_router))) =
-                    prev_id_and_endpoint
+                if let Some(Endpoint::MethodRouter(prev_method_router)) =
+                    self.routes.get_route(path)
                 {
                     // if we're adding a new `MethodRouter` to a route that already has one just
                     // merge them. This makes `.route("/", get(_)).route("/", post(_))` work
-                    let method_router = prev_method_router.clone().merge(method_router);
-                    self.routes
-                        .route_id_to_endpoint
-                        .insert(route_id, Endpoint::MethodRouter(method_router));
+                    let merged = prev_method_router.clone().merge(method_router);
+                    self.routes.overwrite(path, Endpoint::MethodRouter(merged));
                     return self;
                 } else {
                     Endpoint::MethodRouter(method_router)
@@ -507,189 +485,6 @@ where
             DefaultFallback::Unaltered => Route::new(NotFound).call(req),
             DefaultFallback::Layered(inner) => inner.clone().call(req),
         }
-    }
-}
-
-struct InnerRoutes<T> {
-    route_id_to_endpoint: HashMap<RouteId, T>,
-    node: Arc<Node>,
-}
-
-impl<T> InnerRoutes<T> {
-    fn try_insert<'path>(
-        &mut self,
-        path: &'path str,
-        route: T,
-    ) -> Result<RouteId, InsertError<'path, T>> {
-        if path.ends_with("//") {
-            panic!("double slash: {}", path);
-        }
-
-        let id = RouteId::next();
-
-        let mut node =
-            Arc::try_unwrap(Arc::clone(&self.node)).unwrap_or_else(|node| (*node).clone());
-
-        if let Err(err) = node.insert(path, id) {
-            return Err(InsertError { err, path, route });
-        }
-
-        self.node = Arc::new(node);
-
-        self.route_id_to_endpoint.insert(id, route);
-
-        Ok(id)
-    }
-
-    fn overwrite(&mut self, path: &str, route: T) {
-        match self.try_insert(path, route) {
-            Ok(_) => {}
-            Err(err) => {
-                let route = err.route;
-                let conflicting_path = match err.err {
-                    matchit::InsertError::Conflict { with } => with,
-                    _ => unreachable!(),
-                };
-                let id = self.node.path_to_route_id.get(&*conflicting_path).unwrap();
-                self.route_id_to_endpoint.insert(*id, route);
-            }
-        }
-    }
-
-    fn into_iter(self) -> impl Iterator<Item = (Arc<str>, T)> {
-        self.route_id_to_endpoint
-            .into_iter()
-            .map(move |(route_id, route)| {
-                let path = &self.node.route_id_to_path[&route_id];
-                let path = Arc::clone(path);
-                (path, route)
-            })
-    }
-
-    fn at<'router, 'path>(&'router self, path: &'path str) -> Option<Match<'router, 'path, T>> {
-        let matchit::Match { value: id, params } = self.node.at(path).ok()?;
-
-        #[cfg(feature = "matched-path")]
-        let matched_path = self.node.route_id_to_path.get(id).unwrap();
-
-        let route = self.route_id_to_endpoint.get(id).unwrap();
-
-        Some(Match {
-            params,
-            #[cfg(feature = "matched-path")]
-            matched_path,
-            route,
-        })
-    }
-
-    fn map_routes<F, K>(self, f: F) -> InnerRoutes<K>
-    where
-        F: Fn(T) -> K,
-    {
-        let route_id_to_endpoint = self
-            .route_id_to_endpoint
-            .into_iter()
-            .map(|(id, route)| (id, f(route)))
-            .collect();
-
-        InnerRoutes {
-            route_id_to_endpoint,
-            node: self.node,
-        }
-    }
-}
-
-#[derive(Debug)]
-struct InsertError<'a, T> {
-    err: matchit::InsertError,
-    path: &'a str,
-    route: T,
-}
-
-impl<'a, T> fmt::Display for InsertError<'a, T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "Invalid route {:?}: {}", self.path, self.err)
-    }
-}
-
-struct Match<'router, 'path, T> {
-    params: matchit::Params<'router, 'path>,
-    #[cfg(feature = "matched-path")]
-    matched_path: &'router Arc<str>,
-    route: &'router T,
-}
-
-impl<T> Default for InnerRoutes<T> {
-    fn default() -> Self {
-        Self {
-            route_id_to_endpoint: Default::default(),
-            node: Default::default(),
-        }
-    }
-}
-
-impl<T> Clone for InnerRoutes<T>
-where
-    T: Clone,
-{
-    fn clone(&self) -> Self {
-        Self {
-            route_id_to_endpoint: self.route_id_to_endpoint.clone(),
-            node: Arc::clone(&self.node),
-        }
-    }
-}
-
-impl<T> fmt::Debug for InnerRoutes<T>
-where
-    T: fmt::Debug,
-{
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("InnerRoutes")
-            .field("route_id_to_endpoint", &self.route_id_to_endpoint)
-            .field("node", &self.node)
-            .finish()
-    }
-}
-
-/// Wrapper around `matchit::Router` that supports merging two `Router`s.
-#[derive(Clone, Default)]
-struct Node {
-    inner: matchit::Router<RouteId>,
-    route_id_to_path: HashMap<RouteId, Arc<str>>,
-    path_to_route_id: HashMap<Arc<str>, RouteId>,
-}
-
-impl Node {
-    fn insert(
-        &mut self,
-        path: impl Into<String>,
-        val: RouteId,
-    ) -> Result<(), matchit::InsertError> {
-        let path = path.into();
-
-        self.inner.insert(&path, val)?;
-
-        let shared_path: Arc<str> = path.into();
-        self.route_id_to_path.insert(val, shared_path.clone());
-        self.path_to_route_id.insert(shared_path, val);
-
-        Ok(())
-    }
-
-    fn at<'n, 'p>(
-        &'n self,
-        path: &'p str,
-    ) -> Result<matchit::Match<'n, 'p, &'n RouteId>, MatchError> {
-        self.inner.at(path)
-    }
-}
-
-impl fmt::Debug for Node {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Node")
-            .field("paths", &self.route_id_to_path)
-            .finish()
     }
 }
